lab2/requirements.txt
torch
torchvision
torchaudio
numpy
matplotlib
opencv-python
pillow --only-binary :all:
tqdm
scipy